Changes timePicker behavior
Instead of attempting to send new settings each time the timePicker is tapped, it now adds a 'confirm' button when the new time is different from the current settings. When this 'confirm' button is tapped, the new settings are sent to the hardware.

Also adds 'removeListener' functions to the NotificationListener protocol.